Move Backoffice logic to ModelActorTransaction
And drop the Backoffice class
